requirements.txt
<<<<<<< HEAD
pandas
gspread
oauth2client
flask
=======
Flask
pandas
numpy
joblib
shap
google-api-python-client
google-auth
google-auth-oauthlib
>>>>>>> a69ddf9 (Remplacement complet du projet par la nouvelle version)
